Put tests in one file
